requirements.txt
# Core
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0
peft>=0.10.0          # LoRA / QLoRA
bitsandbytes>=0.43.0  # 4-bit quantization

# Retrieval
rank_bm25>=0.2.2
FlagEmbedding>=1.2.0  # BGE-M3 dense embeddings
faiss-cpu>=1.7.4      # Vector index

# LLM / Inference
ollama>=0.1.8

# Evaluation
scikit-learn>=1.4.0
rouge-score>=0.1.2
evaluate>=0.4.1

# Data & Utils
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0
pyyaml>=6.0.1
jsonlines>=4.0.0

# Notebooks
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
